requirements.txt
fastapi 
uvicorn 
python-dotenv 
google-generativeai 
langchain-google-genai 
langgraph 
langchain 
langchain-community 
chromadb 
gTTS 
requests 
